% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooser.R
\name{chooserInput}
\alias{chooserInput}
\title{Lots of parameters that are hopefully obvious}
\usage{
chooserInput(
  inputId,
  leftLabel,
  rightLabel,
  leftChoices,
  rightChoices,
  size = 5,
  multiple = FALSE
)
}
\arguments{
\item{inputId}{}

\item{leftLabel}{}

\item{rightLabel}{}

\item{leftChoices}{}

\item{rightChoices}{}

\item{size}{}

\item{multiple}{}
}
\value{

}
\description{
Lots of parameters that are hopefully obvious
}
\author{
Garrett G
}
