% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformTable.R
\name{deckTable}
\alias{deckTable}
\title{function to generate a table of drafted cards, with columns corresponding to mana value}
\usage{
deckTable(deck, cmc)
}
\arguments{
\item{deck}{the drafted card vector}

\item{cmc}{the mana value vector}
}
\value{
a matrix of the card names by mana value
}
\description{
function to generate a table of drafted cards, with columns corresponding to mana value
}
\author{
Jason Waddell
}
