% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckTableBuild.R
\name{deckTableBuild}
\alias{deckTableBuild}
\title{function to generate a table of drafted cards with main deck, with columns corresponding to mana value}
\usage{
deckTableBuild(main, deck, cmc)
}
\arguments{
\item{main}{the cards in the main deck}

\item{deck}{the drafted card vector}

\item{cmc}{the mana value vector}
}
\value{
a matrix of the card names by mana value
}
\description{
function to generate a table of drafted cards with main deck, with columns corresponding to mana value
}
\author{
Jason Waddell
}
