% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCubeShiny.R
\name{runCubeShiny}
\alias{runCubeShiny}
\title{Run the shiny application}
\usage{
runCubeShiny(installDependencies = FALSE, ...)
}
\arguments{
\item{installDependencies}{boolean, whether to first install packages listed
in the Suggests field of DESCRIPTION; default value is FALSE}

\item{...}{further arguments that can be passed to \code{\link[shiny]{runApp}}}
}
\value{
no return value
}
\description{
Run the shiny application
}
