% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeDeck.R
\name{summarizeDeck}
\alias{summarizeDeck}
\title{Function to summarize the current deck}
\usage{
summarizeDeck(main, draftedCards, cmc, type)
}
\arguments{
\item{main}{list of cards in the main deck}

\item{draftedCards}{list of cards in the drafted cardpool}

\item{cmc}{converted mana cost vector}

\item{type}{vector of card types (e.g. c("Spell", "Land", ...) )}
}
\value{

}
\description{
Function to summarize the current deck
}
\author{
Jason Waddell
}
